requirements.txt
opencv-python
Pillow
imgaug
numpy
scipy
tqdm
tensorboard
addict
sklearn
pathspec
Cython
numba
progress
matplotlib
easydict
wandb
json
pandas
elementpath
